[metal] Expose BufferMemoryView (#4432)
* [metal] Expose BufferMemoryView

* Auto Format

* fix

* fix
